skickar allt innan security

+++ src/main/java/com/edstrom/WigellMcRental/dto/McDto.java
@@ -3,12 +3,14 @@ package com.edstrom.WigellMcRental.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record McDto(
 
         Long id,
          String model,
          String name,
          String year,
-         Long pricePerDay
+         BigDecimal pricePerDay
 ) {
 }
